Rename `stm32` to `pac`
Also cleanup imports in the examples.
