+++ examples/toggle.rs
@@ -8,13 +8,15 @@
 
 use panic_semihosting as _;
 
+use stm32f3xx_hal as hal;
+
 use cortex_m_rt::entry;
-use stm32f3xx_hal::prelude::*;
-use stm32f3xx_hal::stm32;
+use hal::pac;
+use hal::prelude::*;
 
 #[entry]
 fn main() -> ! {
-    let dp = stm32::Peripherals::take().unwrap();
+    let dp = pac::Peripherals::take().unwrap();
 
     let mut rcc = dp.RCC.constrain();
     let mut gpioe = dp.GPIOE.split(&mut rcc.ahb);
