+++ src/flash.rs
@@ -1,6 +1,6 @@
 //! Flash memory
 
-use crate::stm32::{flash, FLASH};
+use crate::pac::{flash, FLASH};
 
 /// Extension trait to constrain the FLASH peripheral
 pub trait FlashExt {
