+++ src/timer.rs
@@ -11,7 +11,7 @@ use crate::hal::timer::{CountDown, Periodic};
     feature = "stm32f378",
     feature = "stm32f398",
 ))]
-use crate::stm32::TIM1;
+use crate::pac::TIM1;
 #[cfg(any(
     feature = "stm32f303",
     feature = "stm32f318",
@@ -20,7 +20,7 @@ use crate::stm32::TIM1;
     feature = "stm32f378",
     feature = "stm32f398"
 ))]
-use crate::stm32::TIM20;
+use crate::pac::TIM20;
 #[cfg(any(
     feature = "stm32f303",
     feature = "stm32f318",
@@ -30,7 +30,7 @@ use crate::stm32::TIM20;
     feature = "stm32f378",
     feature = "stm32f398"
 ))]
-use crate::stm32::TIM4;
+use crate::pac::TIM4;
 #[cfg(any(
     feature = "stm32f303",
     feature = "stm32f318",
@@ -39,10 +39,10 @@ use crate::stm32::TIM4;
     feature = "stm32f378",
     feature = "stm32f398",
 ))]
-use crate::stm32::TIM8;
+use crate::pac::TIM8;
 #[cfg(feature = "stm32f373")]
-use crate::stm32::{TIM12, TIM13, TIM14, TIM18, TIM19, TIM5};
-use crate::stm32::{TIM15, TIM16, TIM17, TIM2, TIM6};
+use crate::pac::{TIM12, TIM13, TIM14, TIM18, TIM19, TIM5};
+use crate::pac::{TIM15, TIM16, TIM17, TIM2, TIM6};
 #[cfg(any(
     feature = "stm32f303",
     feature = "stm32f318",
@@ -53,7 +53,7 @@ use crate::stm32::{TIM15, TIM16, TIM17, TIM2, TIM6};
     feature = "stm32f378",
     feature = "stm32f398"
 ))]
-use crate::stm32::{TIM3, TIM7};
+use crate::pac::{TIM3, TIM7};
 
 use cast::{u16, u32};
 use nb;
